feature/currency_job
