+++ api/src/main/java/com/github/bablo_org/bablo_project/api/controller/CurrencyController.java
@@ -2,14 +2,13 @@ package com.github.bablo_org.bablo_project.api.controller;
 
 import java.util.List;
 
+import com.github.bablo_org.bablo_project.api.client.CurrencyInfoClient;
+import com.github.bablo_org.bablo_project.api.client.CurrencyRatesClient;
 import com.github.bablo_org.bablo_project.api.model.Currency;
 import com.github.bablo_org.bablo_project.api.service.CurrencyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -18,9 +17,24 @@ public class CurrencyController extends BaseController {
 
     private final CurrencyService service;
 
+    private final CurrencyRatesClient apiRates;
+
+    private final CurrencyInfoClient apiInfo;
+
     @GetMapping
     @ResponseBody
     List<Currency> getAll() {
         return service.getAll();
+    }
+
+
+    @PutMapping("/updateRates")
+    void updateRates(){
+        service.updateRates(apiRates.getRates());
+    }
+
+    @PutMapping("/updateInfo")
+    void updateInfo(){
+        service.updateInfo(apiInfo.getInfo());
     }
 }
