+++ api/src/main/java/com/github/bablo_org/bablo_project/api/model/CurrencyAPI/CurrencyInfo.java
@@ -1,25 +1,26 @@
-package com.github.bablo_org.bablo_project.api.model.exchangeRate;
+package com.github.bablo_org.bablo_project.api.model.CurrencyAPI;
 
 import lombok.Value;
 
-import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
 import static java.util.Optional.ofNullable;
 
+
 @Value
-public class CurrencyRate {
+public class CurrencyInfo {
     String id;
-    Double rate;
-    Date updated;
+    String name;
+    String symbol;
 
     public Map<String, Object> toMap() {
         Map<String, Object> map = new HashMap<>();
 
         ofNullable(getId()).ifPresent(v -> map.put("id", v));
-        ofNullable(getRate()).ifPresent(v -> map.put("rate", v));
-        ofNullable(getUpdated()).ifPresent(v -> map.put("updated", v));
+        ofNullable(getName()).ifPresent(v -> map.put("name", v));
+        ofNullable(getSymbol()).ifPresent(v -> map.put("symbol", v));
         return map;
     }
+
 }
