+++ api/src/main/java/com/github/bablo_org/bablo_project/api/model/CurrencyAPI/CurrencyRates.java
@@ -1,24 +1,25 @@
-package com.github.bablo_org.bablo_project.api.model.exchangeRate;
+package com.github.bablo_org.bablo_project.api.model.CurrencyAPI;
 
 import lombok.Value;
 
+import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
 import static java.util.Optional.ofNullable;
 
-
 @Value
-public class CurrencyInfo {
+public class CurrencyRates {
     String id;
-    String name;
+    Double rate;
+    Date updated;
 
     public Map<String, Object> toMap() {
         Map<String, Object> map = new HashMap<>();
 
         ofNullable(getId()).ifPresent(v -> map.put("id", v));
-        ofNullable(getName()).ifPresent(v -> map.put("name", v));
+        ofNullable(getRate()).ifPresent(v -> map.put("rate", v));
+        ofNullable(getUpdated()).ifPresent(v -> map.put("updated", v));
         return map;
     }
-
 }
