+++ api/src/main/java/com/github/bablo_org/bablo_project/api/model/CurrencyAPI/CurrencyRatesToken.java
@@ -1,4 +1,4 @@
-package com.github.bablo_org.bablo_project.api.model.exchangeRate;
+package com.github.bablo_org.bablo_project.api.model.CurrencyAPI;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ExchangeRatesToken {
+public class CurrencyRatesToken {
     private String token;
 }
