+++ api/src/main/java/com/github/bablo_org/bablo_project/api/utils/Resource.java
@@ -1,4 +1,4 @@
-package com.github.bablo_org.bablo_project.api.job;
+package com.github.bablo_org.bablo_project.api.utils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
