+++ api/src/main/resources/application.properties
@@ -5,3 +5,11 @@ logging.level.root=info
 spring.mvc.dispatch-options-request=true
 spring.servlet.multipart.maxFileSize=10MB
 spring.servlet.multipart.maxRequestSize=10MB
+
+filename.currency-names=currency-names.json
+filename.currency-symbols=currency-symbols.json
+
+api.exchange-rate=https://v6.exchangerate-api.com/v6/%s/latest/USD
+api.exchange-rate.fieldname=conversion_rates
+
+firebase.collection_name.currencies=currencies
