+++ api/src/main/resources/currency-symbols.json
@@ -1,0 +1,164 @@
+{
+  "AED": "د.إ",
+  "AFN": "؋",
+  "ALL": "Lek",
+  "AMD": "֏",
+  "ANG": "ƒ",
+  "AOA": "Kz",
+  "ARS": "$",
+  "AUD": "$",
+  "AWG": "ƒ",
+  "AZN": "₼",
+  "BAM": "KM",
+  "BBD": "$",
+  "BDT": "Tk",
+  "BGN": "лв",
+  "BHD": "BD",
+  "BIF": "Franc",
+  "BMD": "$",
+  "BND": "$",
+  "BOB": "$b",
+  "BRL": "R$",
+  "BSD": "$",
+  "BTN": "Ngultrum",
+  "BWP": "P",
+  "BYN": "Br",
+  "BZD": "BZ$",
+  "CAD": "$",
+  "CDF": "Franc",
+  "CHF": "CHF",
+  "CLP": "$",
+  "CNY": "¥",
+  "COP": "$",
+  "CRC": "₡",
+  "CUP": "₱",
+  "CVE": "$",
+  "CZK": "Kč",
+  "DJF": "Franc",
+  "DKK": "kr",
+  "DOP": "RD$",
+  "DZD": "DA",
+  "EGP": "£",
+  "ERN": "Nkf",
+  "ETB": "Br",
+  "EUR": "€",
+  "FJD": "$",
+  "FKP": "£",
+  "FOK": "kr",
+  "GBP": "£",
+  "GEL": "ლ",
+  "GGP": "£",
+  "GHS": "¢",
+  "GIP": "£",
+  "GMD": "D",
+  "GNF": "FG",
+  "GTQ": "Q",
+  "GYD": "$",
+  "HKD": "$",
+  "HNL": "L",
+  "HRK": "kn",
+  "HTG": "G",
+  "HUF": "Ft",
+  "IDR": "Rp",
+  "ILS": "₪",
+  "IMP": "£",
+  "INR": "₹",
+  "IQD": "د.ع",
+  "IRR": "﷼",
+  "ISK": "kr",
+  "JEP": "£",
+  "JMD": "J$",
+  "JOD": "د.أ",
+  "JPY": "¥",
+  "KES": "KSh",
+  "KGS": "лв",
+  "KHR": "៛",
+  "KID": "$",
+  "KMF": "CF",
+  "KRW": "₩",
+  "KWD": "ك",
+  "KYD": "$",
+  "KZT": "лв",
+  "LAK": "₭",
+  "LBP": "£",
+  "LKR": "₨",
+  "LRD": "$",
+  "LSL": "L",
+  "LYD": "LD",
+  "MAD": "درهم",
+  "MDL": "L",
+  "MGA": "Ar",
+  "MKD": "ден",
+  "MMK": "K",
+  "MNT": "₮",
+  "MOP": "MOP$",
+  "MRU": "UM",
+  "MUR": "₨",
+  "MVR": "Rf",
+  "MWK": "MK",
+  "MXN": "$",
+  "MYR": "RM",
+  "MZN": "MT",
+  "NAD": "$",
+  "NGN": "₦",
+  "NIO": "C$",
+  "NOK": "kr",
+  "NPR": "₨",
+  "NZD": "$",
+  "OMR": "﷼",
+  "PAB": "B/.",
+  "PEN": "S/.",
+  "PGK": "K",
+  "PHP": "₱",
+  "PKR": "₨",
+  "PLN": "zł",
+  "PYG": "Gs",
+  "QAR": "﷼",
+  "RON": "lei",
+  "RSD": "Дин.",
+  "RUB": "₽",
+  "RWF": "R₣",
+  "SAR": "﷼",
+  "SBD": "$",
+  "SCR": "₨",
+  "SDG": "LS",
+  "SEK": "kr",
+  "SGD": "$",
+  "SHP": "£",
+  "SLE": "LS",
+  "SLL": "Le",
+  "SOS": "S",
+  "SRD": "$",
+  "SSP": "LS",
+  "STN": "Db",
+  "SYP": "£",
+  "SZL": "E",
+  "THB": "฿",
+  "TJS": "SM",
+  "TMT": "t",
+  "TND": "TD",
+  "TOP": "T$",
+  "TRY": "₺",
+  "TTD": "TT$",
+  "TVD": "$",
+  "TWD": "NT$",
+  "TZS": "TSh",
+  "UAH": "₴",
+  "UGX": "UGX",
+  "USD": "$",
+  "UYU": "$U",
+  "UZS": "лв",
+  "VES": "Bs. S",
+  "VND": "₫",
+  "VUV": "VT",
+  "WST": "$",
+  "XAF": "XAF",
+  "XCD": "$",
+  "XDR": "SDR",
+  "XOF": "XOF",
+  "XPF": "F",
+  "YER": "﷼",
+  "ZAR": "R",
+  "ZMW": "ZK",
+  "ZWL": "$"
+}
